% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhdensities.R
\name{JHDensities}
\alias{JHDensities}
\alias{djh}
\alias{pjh}
\alias{rjh}
\title{The Jones-Henderson Example Densities}
\usage{
djh(x, index, log = FALSE)

pjh(q, index, lower.tail = TRUE, log.p = FALSE)

rjh(n, index)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{index}{an integer between \code{1} and \code{16} inclusive.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to
be the number required.}
}
\value{
\code{djh} gives the density, \code{pjh} gives the distribution function, and
\code{rjh} generates random deviates.
}
\description{
Density, distribution function, and random generation for the sixteen
Jones-Henderson example densities on the unit interval.
}
\references{
Jones, M. C., & Henderson, D. A. (2007). Miscellanea kernel-type
density estimation on the unit interval. Biometrika, 94(4), 977-984.\preformatted{Jones, M. C., & Henderson, D. A. (2007). Kernel-Type Density Estimation
on the Unit Interval,” Statistics Group Technical Report, the Open
University

Geenens, G. (2014). Probit transformation for kernel density estimation
on the unit interval. Journal of the American Statistical Association,
109(505), 346-358.
}
}
